refactor: make tests pass when redis feature is disabled in store
