--- crates/interledger-store/tests/common/mod.rs
@@ -1,9 +1,0 @@
-mod fixtures;
-mod redis_helpers;
-mod store_helpers;
-
-pub use fixtures::*;
-pub use futures::Future;
-pub use interledger_store::redis::*;
-pub use redis_helpers::*;
-pub use store_helpers::*;
