+++ crates/interledger-store/tests/redis/settlement_test.rs
@@ -1,8 +1,6 @@
-mod common;
-
+use super::store_helpers::*;
 use bytes::Bytes;
-use common::*;
-use futures::future::join_all;
+use futures::future::{join_all, Future};
 use http::StatusCode;
 use interledger_api::NodeStore;
